% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-remotes.R
\name{as.package.remotes}
\alias{as.package.remotes}
\title{A simple alternative to \code{devtools::as.package}}
\usage{
as.package.remotes(x)
}
\arguments{
\item{x}{A package object to coerce}
}
\description{
Functionally identical to \code{devtools}' \code{as.package}, but without interactive
options for package creation.
}
\details{
Function required for communicating with the \code{remotes} package interface
}
\note{
Code inspired by \code{devtools} \code{load_pkg_description} with very minor
edits to further reduce \code{devtools} dependencies.
}
\keyword{internal}
