% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deps.R
\name{as_pkg_dependencies}
\alias{as_pkg_dependencies}
\title{Convert a value to a set of dependency types}
\usage{
as_pkg_dependencies(x)
}
\arguments{
\item{x}{A \code{logical} scalar, \code{character} string of \code{"all"}, \code{"most"},
\code{"hard"} or \code{"soft"}, \code{NA} or a vector of dependency types.}
}
\description{
Implements conventions established by both \code{\link[tools:package_dependencies]{tools::package_dependencies()}}
and \code{\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies()}}, following
\code{\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies()}} return type structure of a list
including \verb{$direct} and \verb{$indirect} dependency types. Reimplemented to
avoid dependence on \code{\link{pkgdepends}} compilation requirements.
}
\note{
locally defined and bespoke dispatch system to avoid registering
methods when loaded in combination with \code{pkgdepends}
}
\keyword{internal}
