% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{check_task}
\alias{check_task}
\title{Create a task to run \verb{R CMD check}}
\usage{
check_task(build_args = NULL, args = NULL, env = NULL, ...)
}
\arguments{
\item{build_args}{Character vector of arguments to pass to \verb{R CMD build}.
Pass each argument as a single element of this character vector (do not use
spaces to delimit arguments like you would in the shell). For example,
\code{build_args = c("--force", "--keep-empty-dirs")} is a correct usage and
\code{build_args = "--force --keep-empty-dirs"} is incorrect.}

\item{args}{Character vector of arguments to pass to \verb{R CMD check}. Pass each
argument as a single element of this character vector (do not use spaces to
delimit arguments like you would in the shell). For example, to skip
running of examples and tests, use \code{args = c("--no-examples", "--no-tests")} and not \code{args = "--no-examples --no-tests"}. (Note that
instead of the \code{--output} option you should use the \code{check_dir} argument,
because  \code{--output} cannot deal with spaces and other special characters on
Windows.)}

\item{env}{A named character vector, extra environment variables to
set in the check process.}

\item{...}{
  Arguments passed on to \code{\link{task}}
  \describe{
    \item{\code{.subclass}}{Additional subclasses.}
  }}
}
\description{
Create a task to run \verb{R CMD check}
}
\seealso{
Other tasks:
\code{\link[=install_task]{install_task()}},
\code{\link[=meta_task]{meta_task()}},
\code{\link[=task]{task()}}
}
\concept{tasks}
