% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{checked-package}
\alias{checked}
\alias{checked-package}
\title{checked: Systematically Run R CMD Checks}
\description{
Systematically Run R checks against multiple packages. Checks are run in parallel with strategies to minimize dependency installation. Provides out of the box interface for running reverse dependency check.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Genentech.github.io/checked/}
  \item \url{https://github.com/Genentech/checked}
  \item Report bugs at \url{https://github.com/Genentech/checked/issues}
}

}
\author{
\strong{Maintainer}: Szymon Maksymiuk \email{sz.maksymiuk@gmail.com} (\href{https://orcid.org/0000-0002-3120-1601}{ORCID})

Authors:
\itemize{
  \item Szymon Maksymiuk \email{sz.maksymiuk@gmail.com} (\href{https://orcid.org/0000-0002-3120-1601}{ORCID})
  \item Doug Kelkhoff \email{doug.kelkhoff@gmail.com} (\href{https://orcid.org/0009-0003-7845-4061}{ORCID})
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
