% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_process.R
\name{checks_capture}
\alias{checks_capture}
\title{Parse R CMD checks from a partial check output string}
\usage{
checks_capture(x)
}
\arguments{
\item{x}{A string, compsoed of any subsection of R CMD check console output}
}
\value{
A matrix of matches and capture groups "check" and "status"
("OK", "NONE", "NOTE", "WARNING" or "ERROR").
}
\description{
Parse R CMD checks from a partial check output string
}
\examples{
check_output <- "
* checking check one ... OK
* checking check two ... NOTE
* checking tests ...
  Running test_abc.R
  Running test_xyz.R
 NONE
* checking check three ... WARNING
* ch
"

checks_capture(check_output)

}
