% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_process.R
\name{checks_simplify}
\alias{checks_simplify}
\title{Simplify Captures into Vector}
\usage{
checks_simplify(x)
}
\arguments{
\item{x}{Matrix of regex captures as produced by \link{checks_capture}.}
}
\value{
A vector of check status, with names indicating the check
}
\description{
Simplify Captures into Vector
}
