% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli}
\alias{cli}
\alias{cli_table_row}
\alias{cli_theme}
\title{Internal Utilities for Command-line Output}
\usage{
cli_table_row(
  status = "",
  ok = "OK",
  notes = "N",
  warnings = "W",
  errors = "E",
  msg = "",
  style = c("row", "title", "header"),
  symbols = list(bar = "│")
)

cli_theme(..., .envir = parent.frame())
}
\arguments{
\item{status, ok, notes, warnings, errors}{\code{character[1L]} A value to include
in the respective columns of the table. Will be coerced to \code{character}
if another type is provided.}

\item{msg}{\code{character[1L]} A message to include to the right of the  table
row entry.}

\item{..., .envir}{Additional arguments passed to \code{\link[cli:cli_div]{cli::cli_div()}}}
}
\description{
Various helper functions for consistent cli output, including theming and
formatting.
}
\keyword{internal}
