% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{custom_install_task}
\alias{custom_install_task}
\title{Create a custom install task}
\usage{
custom_install_task(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=install_task]{install_task}}
  \describe{
    \item{\code{lib}}{Any object that can be passed to \code{\link[=lib]{lib()}} to generate a library
path.}
    \item{\code{type}}{character, indicating the type of package to download and
    install.  Will be \code{"source"} except on Windows and some macOS
    builds: see the section on \sQuote{Binary packages} for those.
  }
    \item{\code{INSTALL_opts}}{
    an optional character vector of additional option(s) to be passed to
    \command{R CMD INSTALL} for a source package install.  E.g.,
    \code{c("--html", "--no-multiarch", "--no-test-load")}.

    Can also be a named list of character vectors to be used as
    additional options, with names the respective package names.
  }
  }}
}
\description{
Create a custom install task
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{install_task}()},
\code{\link{revdep_check_task}()},
\code{\link{task}()},
\code{\link{with_subtasks}()}
}
\concept{tasks}
