% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spec.R
\name{custom_install_task_spec}
\alias{custom_install_task_spec}
\title{Create a custom install task}
\usage{
custom_install_task_spec(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=install_task_spec]{install_task_spec}}
  \describe{
    \item{\code{type}}{character, indicating the type of package to download and
    install.  Will be \code{"source"} except on Windows and some macOS
    builds: see the section on \sQuote{Binary packages} for those.
  }
    \item{\code{INSTALL_opts}}{
    an optional character vector of additional option(s) to be passed to
    \command{R CMD INSTALL} for a source package install.  E.g.,
    \code{c("--html", "--no-multiarch", "--no-test-load")}.

    Can also be a named list of character vectors to be used as
    additional options, with names the respective package names.
  }
  }}
}
\description{
Create a custom install task
}
\seealso{
Other tasks: 
\code{\link{check_task_spec}()},
\code{\link{checked-task-df}},
\code{\link{install_task_spec}()},
\code{\link{rev_dep_check_tasks_df}()},
\code{\link{revdep_check_task_spec}()},
\code{\link{source_check_tasks_df}()},
\code{\link{task_spec}()}
}
\concept{tasks}
