% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{DB_COLNAMES}
\alias{DB_COLNAMES}
\title{Available packages database dependencies columns}
\format{
An object of class \code{character} of length 6.
}
\usage{
DB_COLNAMES
}
\description{
Available packages database dependencies columns
}
\keyword{internal}
