% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{DEP}
\alias{DEP}
\title{Dependencies categories}
\format{
An object of class \code{enum} (inherits from \code{factor}) of length 5.
}
\usage{
DEP
}
\description{
Dependencies categories
}
\keyword{internal}
