% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_graph.R
\name{dep_graph_create}
\alias{dep_graph_create}
\title{Create Dependency Graph}
\usage{
dep_graph_create(pkg, ...)
}
\arguments{
\item{pkg, ...}{Passed to \link[miniCRAN:makeDepGraph]{miniCRAN::makeDepGraph}}
}
\value{
A dependency graph with vertex attributes "root" (a logical value
indicating whether the package as one of the roots used to create the
graph), "status" (installation status) and "order" (installation order).
}
\description{
Create Dependency Graph
}
