% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_graph.R
\name{dep_graph_neighborhoods}
\alias{dep_graph_neighborhoods}
\title{Find Dependency Neighborhood}
\usage{
dep_graph_neighborhoods(g, names)
}
\arguments{
\item{g}{A dependency graph, as produced with \code{\link[=dep_graph_create]{dep_graph_create()}}}

\item{names}{Names of packages whose neighborhoods should be calculcated.}
}
\description{
Find Dependency Neighborhood
}
