% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_graph.R
\name{dep_graph_update_install_order}
\alias{dep_graph_update_install_order}
\title{Add Installation Order to Dependency Graph}
\usage{
dep_graph_update_install_order(g)
}
\arguments{
\item{g}{A \link[igraph:make_graph]{igraph::graph}, expected to contain node attribute \code{type}.}
}
\value{
The \link[igraph:make_graph]{igraph::graph} \code{g}, with additional \code{order} vertex attribute,
indicating the order in which dependencies should be installed.
}
\description{
Add Installation Order to Dependency Graph
}
\note{
Cyclic dependencies are possible. Cyclic dependencies are disallowed for all
hard dependencies on CRAN today, though there have been historical instances
where they appeared on CRAN.

Installation priority is based on:
\enumerate{
\item Total dependency footprint (low to high)
\item Topology (leaf nodes first)
}
}
