% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{DEP_STRONG}
\alias{DEP_STRONG}
\title{Strong dependencies categories}
\format{
An object of class \code{factor} of length 3.
}
\usage{
DEP_STRONG
}
\description{
Strong dependencies categories
}
\keyword{internal}
