% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-igraph.R
\name{graph_dedup_attrs}
\alias{graph_dedup_attrs}
\title{Deduplicate attributes}
\usage{
graph_dedup_attrs(g)
}
\description{
Primarily intended for cleaning up the result of an \code{\link[igraph:union]{igraph::union()}},
which adds duplicated attributes when attributes of the same name exist in
multiple graphs. Searches for suffixes and consolidates attributes,
taking the attribute from the first non-NA value observed.
}
