% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{graph_project}
\alias{graph_project}
\title{Project one graph onto another}
\usage{
graph_project(x, onto, where = c(name = "name"))
}
\description{
Project graph \code{x} onto graph \code{onto}, making a single graph with merged
vertices where the attributes in \code{where} from \code{onto} map to the associated
attributes in \code{x} given by the names of \code{where}. \code{where} \emph{must} contain
an element called \code{name}, used to map vertices.
}
\details{
The resulting graph will have the combined vertices and edges of both graphs.
}
