% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib.character}
\alias{lib.character}
\title{Produce a Library from a Path}
\usage{
\method{lib}{character}(x, ..., lib.loc = c())
}
\arguments{
\item{x}{A \code{character} object}

\item{..., lib.loc}{Additional arguments unused}
}
\description{
Produce a Library from a Path
}
