% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib.lib_loc_default}
\alias{lib.lib_loc_default}
\title{Produce a Default Library Path}
\usage{
\method{lib}{lib_loc_default}(x, ..., lib.loc = .libPaths())
}
\arguments{
\item{x}{A \code{lib_loc_default} object.}

\item{...}{Additional arguments unused}

\item{lib.loc}{Library paths, defaulting to \code{\link[=.libPaths]{.libPaths()}}.}
}
\description{
Produce a Default Library Path
}
