% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib.lib_loc_isolated}
\alias{lib.lib_loc_isolated}
\title{Produce an Isolated Library Path}
\usage{
\method{lib}{lib_loc_isolated}(x, name = rlang::hash(runif(1)), lib.root = tempdir(), ...)
}
\arguments{
\item{x}{A \code{lib_loc_isolated} object.}

\item{name}{A name for the library, defaults to a random hash.}

\item{lib.root}{A root directory for the isolated library.}

\item{...}{Additional arguments unused}
}
\description{
Produce an Isolated Library Path
}
