% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib_loc}
\alias{lib_loc}
\title{Make a Library Location}
\usage{
lib_loc(x = list(), .class = c())
}
\arguments{
\item{x}{Any additional metadata that can be prescribed when describing
a library location and used when building the path. Defaults to an empty
\code{\link[=list]{list()}}.}

\item{.class}{An optional subclass, used primarily for dispatch.}
}
\description{
A description of where packages should be installed. This object provides
necessary information to determine where a package should be installed.
}
