% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{library_task}
\alias{library_task}
\title{Specify a library install}
\usage{
library_task(packages, loc = lib_loc_default(), ...)
}
\description{
A declarative task that specifies a set of packages that should be installed
to a given location. This task is typically used during planning of checks,
but does not spawn a process. It is used primarily for organizing the
library precedence for check tasks.
}
