% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_process.R
\name{message_possible_isolation_problems}
\alias{message_possible_isolation_problems}
\title{Message if isolation impossible}
\usage{
message_possible_isolation_problems()
}
\description{
If \emph{R_CHECK_SUGGESTS_ONLY} is set to true, R CMD check will isolate package
installation into temporary directory for running tests and examples. However,
isolation is not applied to dependencies installed in the R_HOME library.
The function informs about possible isolation problem if there are any non
base/recommended packages installed in the .Library (R_HOME).
}
