% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{meta_task}
\alias{meta_task}
\title{Construct a 'Meta' Task}
\usage{
meta_task(..., .subclass = NULL)
}
\arguments{
\item{...}{Objects passed to specified class functions}

\item{.subclass}{character name of the subclass. It will be appended with
"_meta" suffix.}
}
\description{
Meta tasks are tasks which are not intended to perform computation. They
exist simply to provide relationships among computational tasks.
}
\seealso{
Other tasks:
\code{\link[=check_task]{check_task()}},
\code{\link[=install_task]{install_task()}},
\code{\link[=task]{task()}}
}
\concept{tasks}
