% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{new_check_design}
\alias{new_check_design}
\alias{new_rev_dep_check_design}
\title{Create a New Check Design}
\usage{
new_check_design(...)

new_rev_dep_check_design(x, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=new_check_design]{new_check_design()}}}

\item{x}{A file path, passed to \code{\link[=rev_dep_check_tasks_df]{rev_dep_check_tasks_df()}}}
}
\description{
Instantiate a check design from a path or directory.
}
