% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{ORIGINS}
\alias{ORIGINS}
\title{Origins priority order}
\format{
An object of class \code{enum} (inherits from \code{factor}) of length 0.
}
\usage{
ORIGINS
}
\description{
Origins priority order
}
\keyword{internal}
