% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_spec.R
\name{package_spec}
\alias{package_spec}
\alias{package_spec_source}
\alias{package_spec_archive_source}
\title{Package specification}
\usage{
package_spec(name = NULL, repos = NULL)

package_spec_source(path = NULL, ...)

package_spec_archive_source(path = NULL, ...)
}
\arguments{
\item{name}{name of the package.}

\item{repos}{repository where package with given name should identified.}

\item{path}{path to the source of the package (either bundled or not). URLs
are acceptable.}

\item{...}{parameters passed to downstream constructors}
}
\description{
Create package specification list which consists of all the details required
to identify and acquire source of the package.
}
