% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{packages_edges}
\alias{packages_edges}
\title{Produce Graph Edges from Packages Index}
\usage{
packages_edges(ap)
}
\arguments{
\item{ap}{\code{matrix}, as produced by \code{\link[utils:available.packages]{utils::available.packages()}}}
}
\value{
\code{data.frame} with columns \code{package}, \code{dep} and \code{type} and
one row for each dependency relationship.
}
\description{
Produce Graph Edges from Packages Index
}
