% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paths.R
\name{path_parts}
\alias{path_parts}
\title{Split a Filepath into Parts}
\usage{
path_parts(x)
}
\arguments{
\item{x}{A \code{character(1L)} or \code{filepath}}
}
\value{
A \code{character} vector of path parts
}
\description{
Split a Filepath into Parts
}
\keyword{internal}
