% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_origin.R
\name{pkg_origin}
\alias{pkg_origin}
\alias{pkg_origin_repo}
\alias{pkg_origin_base}
\alias{pkg_origin_unknown}
\alias{pkg_origin_local}
\alias{pkg_origin_remote}
\alias{pkg_origin_archive}
\title{Package specification}
\usage{
pkg_origin(package, ..., .class = c())

pkg_origin_repo(package, repos, ...)

pkg_origin_base(package, ...)

pkg_origin_unknown(package, ...)

pkg_origin_local(path = NULL, ...)

pkg_origin_remote(remote = NULL, ...)

pkg_origin_archive(path = NULL, ...)
}
\arguments{
\item{package}{name of the package.}

\item{...}{parameters passed to downstream constructors.}

\item{.class}{Additional subclasses.}

\item{repos}{repository where package with given name should identified.}

\item{path}{path to the source of the package (either bundled or not). URLs
are acceptable.}

\item{remote}{remote object from the \code{remotes} package used to identify
non-standard packages.}
}
\description{
Create package specification list which consists of all the details required
to identify and acquire source of the package.
}
\seealso{
Other specs:
\code{\link[=lib_path]{lib_path()}}
}
\concept{specs}
