% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{PLAN}
\alias{PLAN}
\title{Check execution status categories}
\format{
An object of class \code{list} of length 2.
}
\usage{
PLAN
}
\description{
Check execution status categories
}
\keyword{internal}
