% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_checks}
\alias{plan_checks}
\title{Plan multiple R CMD checks from source package paths}
\usage{
plan_checks(path)
}
\arguments{
\item{path}{path to the package source. Can be either a single source
code directory or a directory containing multiple package source code
directories.}
}
\value{
The check schedule \code{data.frame} with the following columns:
\itemize{
\item \code{alias}: The alias of the check to run. It also serves the purpose of
providing a unique identifier and node name in the task graph.
\item \code{version}: Version of the package to be checked.
\item \code{package}: Object that inherits from \code{\link[=check_task]{check_task()}}.
Defines how package to be checked can be acquired.
\item \code{custom}:  Object that inherits from \code{\link[=custom_install_task]{custom_install_task()}}.
Defines custom package, for instance only available from local source, that
should be installed before checking the package.
}
}
\description{
Plans are pre-specified sets of checks. Plans are simple \code{data.frame}s
where each row defines a package for which \verb{R CMD check}
should be run.
}
\seealso{
Other plan: 
\code{\link{plan}},
\code{\link{plan_rev_dep_checks}()}
}
\concept{plan}
