% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_local_install}
\alias{plan_local_install}
\title{Plan source package installation}
\usage{
plan_local_install(
  package,
  repos = getOption("repos"),
  remotes_dependencies = TRUE,
  INSTALL_opts = c()
)
}
\arguments{
\item{package}{A path to package source.}

\item{repos}{repository used to identify packages when name is provided.}

\item{remotes_dependencies}{A vector of length one or a named list.
Compatible with \code{\link{as_pkg_dependencies}}. Used to filter out remotes
dependencies.}

\item{INSTALL_opts}{Options to set while the root package is being installed.
Check \code{\link[utils:install.packages]{utils::install.packages}} for details.}
}
\description{
Generates a plan for running installing a package from source.
}
\seealso{
Other plan:
\code{\link[=plan_local_checks]{plan_local_checks()}},
\code{\link[=plan_rev_dep_checks]{plan_rev_dep_checks()}}
}
\concept{plan}
