% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcheck_process.R
\docType{data}
\name{RE_CHECK}
\alias{RE_CHECK}
\title{Regular Expression for Parsing R CMD check checks}
\format{
An object of class \code{character} of length 1.
}
\usage{
RE_CHECK
}
\description{
Regular Expression for Parsing R CMD check checks
}
\keyword{datasets}
