% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\docType{data}
\name{RELATION}
\alias{RELATION}
\title{Relation types}
\format{
An object of class \code{enum} (inherits from \code{factor}) of length 2.
}
\usage{
RELATION
}
\description{
A flag for edges that articulate different relations between nodes.
}
\keyword{internal}
