% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{reporters}
\alias{reporters}
\alias{reporter_ansi_tty}
\alias{reporter_basic_tty}
\alias{reporter_default}
\title{Check Design Runner Reporters}
\usage{
reporter_ansi_tty()

reporter_basic_tty()

reporter_default()
}
\description{
Reporters are used to configure how output is communicated while running
a \code{\link{check_design}}. They range from glossy command-line tools intended for
displaying progress in an interactive R session, to line-feed logs which
may be better suited for automated execution, such as in continuous
itegration.
}
\details{
\subsection{\code{\link[=reporter_default]{reporter_default()}}}{

Automatically chooses an appropriate reporter based on the calling context.
}

\subsection{\code{\link[=reporter_ansi_tty]{reporter_ansi_tty()}}}{

Highly dynamic output for fully capable terminals. Requires multi-line
dynamic output, which may not be available in editors that that present a
terminal as a web component.
}

\subsection{\code{\link[=reporter_basic_tty]{reporter_basic_tty()}}}{

A line-feed reporter presenting output one line at a time, providing
a reporter with minimal assumptions about terminal capabilities.
}
}
