% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results_to_df}
\alias{results_to_df}
\title{Summarize checked results as data.frame}
\usage{
results_to_df(results, ...)
}
\arguments{
\item{results}{checked_results object or any of the sub-objects.}

\item{...}{other prameters passed to downstream functions.}
}
\description{
Turns checked results into a list of data.frams, one for each meta root task.
Such form makes quick results analysis easier by providing a general overview
of identified failures.
}
\seealso{
Other results:
\code{\link[=print.checked_results]{print.checked_results()}},
\code{\link[=results]{results()}}
}
\concept{results}
