% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results_to_file}
\alias{results_to_file}
\title{Results to file}
\usage{
results_to_file(results, file, keep = "all", ...)
}
\arguments{
\item{results}{\code{\link[checked]{results}} object.}

\item{file}{A connection or character path.}

\item{keep}{character vector indicating which packages should be included in the results.
"all" means that all packages are kept. If "issues" then only packages
with issues identified, whereas "potential_issues" stands for keeping
packages with both "issues" and "potential_issues". (Defaults to \code{"all"}, overwritable using option 'checked.results_keep' or environment variable 'R_CHECKED_RESULTS_KEEP')}

\item{...}{other parameters.}
}
\description{
Write \code{checked_results} object to the text file. When converting results
to text, \code{\link[checked]{print.checked_results}} method is used.
}
\seealso{
Other results: 
\code{\link{print.checked_results}()},
\code{\link{results}()}
}
\concept{results}
