% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{rev_dep_attempt_restore}
\alias{rev_dep_attempt_restore}
\title{Attempt to Recover Previous Run State}
\usage{
rev_dep_attempt_restore(path, restore = if (interactive()) NA else TRUE)
}
\description{
Attempt to Recover Previous Run State
}
