% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{revdep_check_task}
\alias{revdep_check_task}
\title{Create a task to run reverse dependency checks}
\usage{
revdep_check_task(revdep, ...)
}
\arguments{
\item{revdep}{character indicating whether the task specification describes
check associated with the development (new) or release (old) version of the
for which reverse dependency check is run.}

\item{...}{Additional parameters passed to \code{\link[=task]{task()}}}
}
\description{
Create a task to run reverse dependency checks
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{custom_install_task}()},
\code{\link{install_task}()},
\code{\link{task}()}
}
\concept{tasks}
