% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spec.R
\name{revdep_check_task_spec}
\alias{revdep_check_task_spec}
\title{Create a task to run reverse dependency checks}
\usage{
revdep_check_task_spec(revdep, ...)
}
\arguments{
\item{revdep}{character indicating whether the task specification describes
check associated with the development (new) or release (old) version of the
for which reverse dependency check is run.}

\item{...}{Additional parameters passed to \code{\link[=task_spec]{task_spec()}}}
}
\description{
Create a task to run reverse dependency checks
}
\seealso{
Other tasks: 
\code{\link{check_task_spec}()},
\code{\link{checked-task-df}},
\code{\link{install_task_spec}()},
\code{\link{rev_dep_check_tasks_df}()},
\code{\link{task_spec}()}
}
\concept{tasks}
