% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_graph.R
\name{revdeps_graph_create}
\alias{revdeps_graph_create}
\title{Create Dependency Graph for Reverse Dependencies}
\usage{
revdeps_graph_create(pkg, db = available.packages(), which = "strong", ...)
}
\arguments{
\item{pkg}{A package whose reverse dependencies should be used as the basis
for the dependency graph.}

\item{db}{character matrix as from \code{\link{available.packages}()}
    (with the default \code{NULL} the results of this call)
    or data frame variants thereof.  Alternatively, a package database
    like the one available from
    \url{https://cran.r-project.org/web/packages/packages.rds}.
  }

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }

\item{...}{
  Arguments passed on to \code{\link[miniCRAN:makeDepGraph]{miniCRAN::makeDepGraph}}
  \describe{
    \item{\code{availPkgs}}{Data frame with an element called \code{package}. The \code{package}
element is a vector of available packages.  Defaults to reading this list
from CRAN, using \code{\link[=available.packages]{available.packages()}}}
    \item{\code{repos}}{URL(s) of the 'contrib' sections of the repositories, e.g.
\code{"https://cran.us.r-project.org"}. Passed to \code{\link[=available.packages]{available.packages()}}}
    \item{\code{type}}{Possible values are (currently) "source", "mac.binary" and
"win.binary": the binary types can be listed and downloaded but not
installed on other platforms.  Passed to \code{\link[=download.packages]{download.packages()}}.}
    \item{\code{suggests}}{If TRUE, retrieves Suggests dependencies (non-recursively)}
    \item{\code{enhances}}{If TRUE, retrieves Enhances dependencies (non-recursively)}
    \item{\code{includeBasePkgs}}{If TRUE, include base R packages in results}
  }}
}
\description{
Discovers reverse dependencies and builds a dependency graph for all reverse
dependency dependencies.
}
