% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{reversecheck-package}
\alias{reversecheck}
\alias{reversecheck-package}
\title{reversecheck: Systematically Run R Checks Against Reverse Dependencies}
\description{
Run R checks against old and new versions of a package to systematically test for any breaking changes imparted on packages which take said package as a dependency (reverse dependencies). Checks are run in parallel with strategies to minimize dependency installation.
}
\author{
\strong{Maintainer}: S M \email{s.m@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Doug Kelkhoff \email{doug.kelkhoff@gmail.com} (\href{https://orcid.org/0009-0003-7845-4061}{ORCID})
}

}
