% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a Series of \verb{R CMD check}s}
\usage{
run(checker, ..., reporter = reporter_default())
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=new_rev_dep_checker]{new_rev_dep_checker()}}}

\item{reporter}{A reporter to provide progress updates. Will default to the
most expressive command-line reporter given your terminal capabilities.}

\item{checks}{\code{character} or \code{checker} If a \code{character} value is
provided, it is first coerced into a \code{checker} using
\code{\link[=new_rev_dep_checker]{new_rev_dep_checker()}}.}
}
\description{
\code{\link[=run]{run()}} provides a generic, and is the central interface for executing
\code{\link{checker}}s. If a path is provided, a new reverse dependency check
plan is generated from the source code path. Otherwise a plan can be
built separately and executed using \code{\link[=run]{run()}}.
}
