% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silent_spinner}
\alias{silent_spinner}
\title{Create a 'cli' Spinner With Suppressed Output}
\usage{
silent_spinner(..., stream = devnull())
}
\arguments{
\item{...}{passed to \link[cli:make_spinner]{cli::make_spinner}}

\item{stream}{passed to \link[cli:make_spinner]{cli::make_spinner}, defaults to a null file device}
}
\value{
A interface similar to a 'cli' spinner, but with suppressed output
}
\description{
'cli' will implicitly push spinner output to various output streams,
affecting the terminal cursor position. To allow for a terminal interface
that has spinners above the last line, this function suppresses the output
and simply returns its frame contents.
}
