% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_task.R
\name{start_task}
\alias{start_task}
\title{Start a new task}
\usage{
start_task(node, g, ...)
}
\arguments{
\item{node}{Node(s) for which libpath should be constructed based on \code{g}}

\item{g}{\code{task_graph} object}

\item{...}{additional params passed to downstream methods}
}
\description{
Starts task based on the \code{task} object encapsulated in the \code{node} taken
from then \code{task_graph} \code{g}. It returns an \code{install_process} or
\code{check_process} \code{R6} object.
}
\keyword{internal}
