% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-enums.R
\name{STATUS}
\alias{STATUS}
\title{Check execution status categories}
\usage{
STATUS
}
\description{
Check execution status categories
}
\seealso{
Other checks:
\code{\link[=check_pkgs]{check_pkgs()}},
\code{\link[=check_rev_deps]{check_rev_deps()}},
\code{\link{checker}},
\code{\link[=new_checker]{new_checker()}}
}
\concept{checks}
