% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg-source.R
\name{sub_aliased_deps}
\alias{sub_aliased_deps}
\title{Substitute Aliased Dependencies String}
\usage{
sub_aliased_deps(deps, aliases)
}
\arguments{
\item{deps}{\code{character(1L)}, as provided by a \code{DESCRIPTION} file's
dependency listing.}

\item{aliases}{(named vector), a mapping of package names to aliases for
substitution.}
}
\value{
\code{deps}, substituting package names with associated aliases.
}
\description{
Substitute Aliased Dependencies String
}
