% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{task_formats}
\alias{task_formats}
\title{Task formatter bindings}
\usage{
task_formats(g = NULL, nodes = V(g), task = NULL, tasks = list(task))
}
\arguments{
\item{g}{task_graph object.}

\item{nodes}{graph nodes to format.}

\item{task}{task to format.}

\item{tasks}{currently unused.}
}
\description{
This bit of code is intended for use with \code{\link[=fmt]{fmt()}}, and allows for us
to layer symbol bindings on top of the environment used for string
interpolation which provide syntactic sugar for common formatting components.
}
\keyword{internal}
