% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_task.R
\name{task_graph_libpaths}
\alias{task_graph_libpaths}
\title{Libpaths from task graph}
\usage{
task_graph_libpaths(g, node = NULL, lib.loc = .libPaths(), output = tempdir())
}
\arguments{
\item{g}{\code{task_graph} object}

\item{node}{Node(s) for which libpath should be constructed based on \code{g}}

\item{lib.loc}{Library paths, defaulting to \code{\link[=.libPaths]{.libPaths()}}.}

\item{output}{Path to the checked output directory}
}
\description{
Function that traverses over the task dependency task to acquire libpaths
for given nodes. It ensures that when runing a node, a libpath is
constructed which has all the required packages on it.
}
\keyword{internal}
