% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{task_graph_sort}
\alias{task_graph_sort}
\title{Sort Task Graph by Strong Dependency Order}
\usage{
task_graph_sort(g)
}
\arguments{
\item{g}{A \link[igraph:make_graph]{igraph::graph}, expected to contain node attribute \code{type}.}
}
\value{
The \link[igraph:make_graph]{igraph::graph} \code{g}, with vertices sorted in preferred
installation order.
}
\description{
Sort Task Graph by Strong Dependency Order
}
\note{
Cyclic dependencies are possible. Cyclic dependencies are disallowed for all
hard dependencies on CRAN today, though there have been historical instances
where they appeared on CRAN.

Installation priority is based on:
\enumerate{
\item Total dependency footprint (low to high)
\item Topology (leaf nodes first)
}
}
