% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{task_graph_which_satisfied}
\alias{task_graph_which_satisfied}
\title{Find the Next Packages Not Dependent on an Unavailable Package}
\usage{
task_graph_which_satisfied(
  g,
  v = igraph::V(g),
  dependencies = TRUE,
  status = STATUS$pending
)
}
\arguments{
\item{g}{A dependency graph, as produced with \code{\link[=dep_graph_create]{dep_graph_create()}}}
}
\value{
The name of the next package to prioritize
}
\description{
While other packages are in progress, ensure that the next selected package
already has its dependencies done.
}
