% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_spec.R
\name{task_spec}
\alias{task_spec}
\title{Task specification}
\usage{
task_spec(
  alias = NULL,
  package_spec = NULL,
  env = options::opt("check_envvars")
)
}
\arguments{
\item{alias}{task alias which also serves as unique identifier of the task.}

\item{package_spec}{\code{\link[checked]{package_spec}} object}

\item{env}{environmental variables to be set in separate process running
specific task.}
}
\description{
Create task specification list which consists of all the details required
to run specific task.
}
\seealso{
Other tasks: 
\code{\link{check_task_spec}()},
\code{\link{checked-task-df}},
\code{\link{custom_install_task_spec}()},
\code{\link{install_task_spec}()},
\code{\link{rev_dep_check_tasks_df}()},
\code{\link{revdep_check_task_spec}()},
\code{\link{source_check_tasks_df}()}
}
\concept{tasks}
