% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_process.R
\name{throttle}
\alias{throttle}
\title{Generate A Rate Limiting Throttle Function}
\usage{
throttle(interval = 0.2)
}
\arguments{
\item{interval}{An interval (in seconds) that is the minimum interval
before \link{throttle} will return \code{TRUE}.}
}
\value{
A throttling function with the provided interval. When called,
returns a logical value indicating whether the throttle interval has
passed (TRUE if the interval has not yet passed).
}
\description{
Generate A Rate Limiting Throttle Function
}
