% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{with_subtasks}
\alias{with_subtasks}
\title{Attach Subtasks to an Existing Task}
\usage{
with_subtasks(task, tasks)
}
\description{
Attach Subtasks to an Existing Task
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{custom_install_task}()},
\code{\link{install_task}()},
\code{\link{revdep_check_task}()},
\code{\link{task}()}
}
\concept{tasks}
