% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{check_task}
\alias{check_task}
\title{Create a task to run \verb{R CMD check}}
\usage{
check_task(args = NULL, build_args = NULL, ...)
}
\arguments{
\item{args}{Character vector of arguments to pass to \verb{R CMD check}. Pass each
argument as a single element of this character vector (do not use spaces to
delimit arguments like you would in the shell). For example, to skip
running of examples and tests, use \code{args = c("--no-examples", "--no-tests")} and not \code{args = "--no-examples --no-tests"}. (Note that
instead of the \code{--output} option you should use the \code{check_dir} argument,
because  \code{--output} cannot deal with spaces and other special characters on
Windows.)}

\item{build_args}{Character vector of arguments to pass to \verb{R CMD build}.
Pass each argument as a single element of this character vector (do not use
spaces to delimit arguments like you would in the shell). For example,
\code{build_args = c("--force", "--keep-empty-dirs")} is a correct usage and
\code{build_args = "--force --keep-empty-dirs"} is incorrect.}

\item{...}{
  Arguments passed on to \code{\link[=task]{task}}
  \describe{
    \item{\code{alias}}{task alias which also serves as unique identifier of the task.}
    \item{\code{package}}{\code{\link[checked]{package}} object}
    \item{\code{env}}{environmental variables to be set in separate process running
specific task.}
  }}
}
\description{
Create a task to run \verb{R CMD check}
}
\seealso{
Other tasks: 
\code{\link{custom_install_task}()},
\code{\link{install_task}()},
\code{\link{revdep_check_task}()},
\code{\link{task}()}
}
\concept{tasks}
