% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_df.R
\name{checked-task-df}
\alias{checked-task-df}
\title{Check schedule data frame}
\arguments{
\item{path}{path to the package source. Can be either a single source
code directory or a directory containing multiple package source code
directories.}

\item{...}{parameters passed to the task specs allowing to customize
subprocesses.}
}
\value{
The check schedule \code{data.frame} with the following columns:
\itemize{
\item \code{alias}: The alias of the check to run. It also serves the purpose of
providing a unique identifier and node name in the task graph.
\item \code{version}: Version of the package to be checked.
\item \code{package}: Object that inherits from \code{\link[=check_task_spec]{check_task_spec()}}.
Defines how package to be checked can be acquired.
\item \code{custom}:  Object that inherits from \code{\link[=custom_install_task_spec]{custom_install_task_spec()}}.
Defines custom package, for instance only available from local source, that
should be installed before checking the package.
}
}
\description{
Create data.frame which each row defines a package for which R CMD check
should be run. Such data.frame is a prerequisite for generating
\code{\link[=check_design]{check_design()}} which orchestrates all the processes
including dependencies installation.
}
\details{
\verb{_tasks_df()} functions generate check task \code{data.frame} for
all source packages specified by the \code{path}. Therefore it accepts it to be
a vector of an arbitrary length.
}
\seealso{
Other tasks: 
\code{\link{check_task_spec}()},
\code{\link{custom_install_task_spec}()},
\code{\link{install_task_spec}()},
\code{\link{rev_dep_check_tasks_df}()},
\code{\link{revdep_check_task_spec}()},
\code{\link{source_check_tasks_df}()},
\code{\link{task_spec}()}
}
\concept{tasks}
