% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_graph.R
\name{dep_graph_which_satisfied}
\alias{dep_graph_which_satisfied}
\alias{dep_graph_which_satisfied_strong}
\alias{dep_graph_which_root_satisfied}
\title{Find the Next Packages Not Dependent on an Unavailable Package}
\usage{
dep_graph_which_satisfied(g, v = igraph::V(g), dependencies = TRUE)

dep_graph_which_satisfied_strong(..., dependencies = "strong")

dep_graph_which_root_satisfied(g, ..., dependencies = "all")
}
\arguments{
\item{g}{A dependency graph, as produced with \code{\link[=dep_graph_create]{dep_graph_create()}}}
}
\value{
The name of the next package to prioritize
}
\description{
While other packages are installing, ensure that the next selected package
already has its dependencies installed.
}
\section{Functions}{
\itemize{
\item \code{dep_graph_which_satisfied_strong()}: List vertices whose strong dependencies are satisfied

\item \code{dep_graph_which_root_satisfied()}: List root vertices whose dependencies are all satisfied

}}
