% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{fmt}
\alias{fmt}
\title{Produce cli output for a task}
\usage{
fmt(..., g, nodes, task = NULL, .envir = parent.frame())
}
\description{
Provided a task, allows for use of a handful of shorthand symbols which will
use the task as a context for formatting task fields.
}
\examples{
task <- install_task(origin = pkg_origin(
  package = "pkg",
  version = package_version("1.2.3"),
  source = "../../Programming/options"
))

fmt(task = task, "{action} {package} ({version}) from {source}")

}
