% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{install_task}
\alias{install_task}
\title{Create a task to install a package and dependencies}
\usage{
install_task(origin, type = getOption("pkgType"), INSTALL_opts = NULL, ...)
}
\arguments{
\item{type}{character, indicating the type of package to download and
    install.  Will be \code{"source"} except on Windows and some macOS
    builds: see the section on \sQuote{Binary packages} for those.
  }

\item{INSTALL_opts}{
    an optional character vector of additional option(s) to be passed to
    \command{R CMD INSTALL} for a source package install.  E.g.,
    \code{c("--html", "--no-multiarch", "--no-test-load")}.

    Can also be a named list of character vectors to be used as
    additional options, with names the respective package names.
  }

\item{...}{Additional parameters passed to \code{\link[=task]{task()}}}
}
\description{
Create a task to install a package and dependencies
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{custom_install_task}()},
\code{\link{revdep_check_task}()},
\code{\link{task}()}
}
\concept{tasks}
