% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib.lib_path_isolated}
\alias{lib.lib_path_isolated}
\title{Produce an Isolated Library Path}
\usage{
\method{lib}{lib_path_isolated}(
  x,
  ...,
  lib.root = tempdir(),
  dir_hash = hash(Sys.time(), n = 8),
  name = ""
)
}
\arguments{
\item{x}{A \code{lib_path_isolated} object.}

\item{...}{Additional arguments unused}

\item{lib.root}{A root directory for the isolated library.}

\item{dir_hash}{unique identifier of the isolated library}
}
\description{
Produce an Isolated Library Path
}
