% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib}
\alias{lib}
\title{Get Library Location}
\usage{
lib(x, ..., lib.loc = c())
}
\arguments{
\item{x}{An object describing a library location}

\item{lib.loc}{A set of library locations, used as defaults for objects
that may make use of them.}
}
\description{
Get Library Location
}
