% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib_path}
\alias{lib_path}
\title{Make a Library Location}
\usage{
lib_path(x, ..., .class = c())
}
\arguments{
\item{x}{A \code{\link[=pkg_origin]{pkg_origin()}} object}

\item{.class}{An optional subclass, used primarily for dispatch.}
}
\description{
A description of where packages should be installed. This object provides
necessary information to determine where a package should be installed.
}
