% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{meta_task}
\alias{meta_task}
\title{Construct a 'Meta' Task}
\usage{
meta_task(..., .subclass = NULL)
}
\description{
Meta tasks are tasks which are not intended to perform computation. They
exist simply to provide relationships among computational tasks.
}
