% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{new_checker}
\alias{new_checker}
\alias{new_rev_dep_checker}
\title{Creating new Check Design Objects}
\usage{
new_checker(...)

new_rev_dep_checker(x, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=new_checker]{new_checker()}}}

\item{x}{A file path, passed to \code{\link[=plan_rev_dep_checks]{plan_rev_dep_checks()}}}
}
\description{
Instantiate a check design from a path or directory.
}
\seealso{
Other checks: 
\code{\link{STATUS}},
\code{\link{check_pkgs}()},
\code{\link{check_rev_deps}()},
\code{\link{checker}}

Other checks: 
\code{\link{STATUS}},
\code{\link{check_pkgs}()},
\code{\link{check_rev_deps}()},
\code{\link{checker}}
}
\concept{checks}
