% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{new_checker}
\alias{new_checker}
\title{Creating new Check Design Objects}
\usage{
new_checker(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=new_check_design]{new_check_design()}}}

\item{x}{A file path, passed to \code{\link[=rev_dep_check_tasks_df]{rev_dep_check_tasks_df()}}}
}
\description{
Instantiate a check design from a path or directory.
}
\seealso{
Other checks: 
\code{\link{check_dev_rev_deps}()},
\code{\link{check_dir}()},
\code{\link{check_pkgs}()},
\code{\link{check_rev_deps}()},
\code{\link{checker}},
\code{\link{new_check_design}}
}
\concept{checks}
