% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{checked Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{checked} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "checked")

options::opt(x, default, env = "checked")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{tty_tick_interval}{\describe{
tty refresh interval when reporting results in milliseconds\item{default: }{\preformatted{0.1}}
\item{option: }{checked.tty_tick_interval}
\item{envvar: }{R_CHECKED_TTY_TICK_INTERVAL (evaluated if possible, raw string otherwise)}
}}

\item{results_error_on}{\describe{
character vector indicating whether R error should be thrown when issues
are discovered when generating results. "never" means that no errors
are thrown. If "issues" then errors are emitted only on issues, whereas
"potential issues" stands for error on both issues and potential issues.\item{default: }{\preformatted{"never"}}
\item{option: }{checked.results_error_on}
\item{envvar: }{R_CHECKED_RESULTS_ERROR_ON (evaluated if possible, raw string otherwise)}
}}

\item{results_keep}{\describe{
character vector indicating which packages should be included in the results.
"all" means that all packages are kept. If "issues" then only packages
with issues identified, whereas "potential_issues" stands for keeping
packages with both "issues" and "potential_issues".\item{default: }{\preformatted{"all"}}
\item{option: }{checked.results_keep}
\item{envvar: }{R_CHECKED_RESULTS_KEEP (evaluated if possible, raw string otherwise)}
}}

\item{restore}{\describe{
\code{logical} indicating whether output directory should be unlinked before
running checks. If \code{FALSE}, an attempt will me made to restore previous
progress from the same \code{output}\item{default: }{\preformatted{NA}}
\item{option: }{checked.restore}
\item{envvar: }{R_CHECKED_RESTORE (evaluated if possible, raw string otherwise)}
}}

\item{default_check_env_variables}{\describe{
\code{logical} indicating whether default R CMD check related enviromental
variables should be appended.\item{default: }{\preformatted{TRUE}}
\item{option: }{checked.default_check_env_variables}
\item{envvar: }{R_CHECKED_DEFAULT_CHECK_ENV_VARIABLES (evaluated if possible, raw string otherwise)}
}}

\item{default_check_build_args}{\describe{
\code{logical} indicating whether default R CMD check build args should be appended.\item{default: }{\preformatted{TRUE}}
\item{option: }{checked.default_check_build_args}
\item{envvar: }{R_CHECKED_DEFAULT_CHECK_BUILD_ARGS (evaluated if possible, raw string otherwise)}
}}

\item{default_check_args}{\describe{
\code{logical} indicating whether default R CMD check args should be appended.\item{default: }{\preformatted{TRUE}}
\item{option: }{checked.default_check_args}
\item{envvar: }{R_CHECKED_DEFAULT_CHECK_ARGS (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
