% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deps.R
\name{pkg_dependencies}
\alias{pkg_dependencies}
\title{Build Package Dependencies Table}
\usage{
pkg_dependencies(
  packages,
  dependencies = TRUE,
  db = available.packages(),
  verbose = FALSE
)
}
\description{
Inspired by \code{tools::package_dependencies}, but with the added benefit
of recording the dependency type and relationships throughout the
dependency tree.
}
