% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_origin.R
\name{pkg_origin}
\alias{pkg_origin}
\alias{pkg_origin_repo}
\alias{pkg_origin_local}
\alias{pkg_origin_archive}
\title{Package specification}
\usage{
pkg_origin(name = NULL)

pkg_origin_repo(repos, ...)

pkg_origin_local(path = NULL, ...)

pkg_origin_archive(path = NULL, ...)
}
\arguments{
\item{name}{name of the package.}

\item{repos}{repository where package with given name should identified.}

\item{...}{parameters passed to downstream constructors}

\item{path}{path to the source of the package (either bundled or not). URLs
are acceptable.}
}
\description{
Create package specification list which consists of all the details required
to identify and acquire source of the package.
}
\concept{specs}
