% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_local_checks}
\alias{plan_local_checks}
\title{Plan R CMD Checks}
\usage{
plan_local_checks(package, repos = getOption("repos"))
}
\arguments{
\item{package}{A path to either package, directory with packages or name
of the package (details)}

\item{repos}{repository used to identify packages when name is provided.}
}
\description{
Generates a plan for running R CMD check for a specified set of packages.
}
\details{
\code{package} parameter has two different allowed values:
\itemize{
\item Package - checked looks for a DESCRIPTION file in the provided path, if
found treats it like a source package.
\item If the specified value does not correspond to a source package, the
parameter is treated as the name and \code{repos} parameter is used to identify
the source.
}
}
\seealso{
Other plan: 
\code{\link{plan_local_install}()},
\code{\link{plan_rev_dep_checks}()}
}
\concept{plan}
