% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_local_install}
\alias{plan_local_install}
\title{Plan source package installation}
\usage{
plan_local_install(package, repos = getOption("repos"))
}
\arguments{
\item{package}{A path to package source.}

\item{repos}{repository used to identify packages when name is provided.}
}
\description{
Generates a plan for running installing a package from source.
}
\seealso{
Other plan: 
\code{\link{plan_local_checks}()},
\code{\link{plan_rev_dep_checks}()}
}
\concept{plan}
