% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_rev_dep_checks}
\alias{plan_rev_dep_checks}
\title{Plan Reverse Dependency Checks}
\usage{
plan_rev_dep_checks(path, repos = getOption("repos"))
}
\arguments{
\item{path}{path to the package source.}

\item{repos}{repository used to identify reverse dependencies.}
}
\description{
Generates a plan for running reverse dependency check for certain
source package. In such case \code{path} should be proivded with a directory
path to the development version of the package and \code{repos} should be a
repository for which reverse dependencies should be identified.
}
\seealso{
Other plan: 
\code{\link{plan_local_checks}()},
\code{\link{plan_local_install}()}
}
\concept{plan}
