% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.checked_results}
\alias{print.checked_results}
\alias{print.rev_dep_dep_results}
\alias{print.local_check_results}
\title{Print checked results}
\usage{
\method{print}{checked_results}(x, ...)

\method{print}{rev_dep_dep_results}(x, ..., name = NULL, keep = options::opt("results_keep"))

\method{print}{local_check_results}(x, ..., name = NULL, keep = options::opt("results_keep"))
}
\arguments{
\item{x}{an object to be printed.}

\item{...}{other parameters.}

\item{name}{character name of the \code{rev_dep_dep} package}

\item{keep}{character vector indicating which packages should be included in the results.
"all" means that all packages are kept. If "issues" then only packages
with issues identified, whereas "potential_issues" stands for keeping
packages with both "issues" and "potential_issues". (Defaults to \code{"all"}, overwritable using option 'checked.results_keep' or environment variable 'R_CHECKED_RESULTS_KEEP')}
}
\description{
Print checked results
}
\seealso{
Other results: 
\code{\link{results}()},
\code{\link{results_to_df}()}
}
\concept{results}
