% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.checked_results}
\alias{print.checked_results}
\alias{print.checked_results_check_task_spec}
\alias{print.checked_results_revdep_check_task_spec}
\title{Print checked results}
\usage{
\method{print}{checked_results}(x, ...)

\method{print}{checked_results_check_task_spec}(
  x,
  keep = Sys.getenv("CHECKED_RESULTS_KEEP", c("all", "issues", "potential_issues")[1]),
  ...
)

\method{print}{checked_results_revdep_check_task_spec}(x, ...)
}
\arguments{
\item{x}{an object to be printed.}

\item{...}{other parameters.}

\item{keep}{character vector indicating which packages should be included
in the results. "all" means that all packages are kept. If "issues" then
only packages with issues identified, whereas "potential_issues" stands for
keeping packages with both "issues" and "potential_issues". Users can set
the default value via env variable \code{CHECKED_RESULTS_KEEP}.}
}
\description{
Print checked results
}
\seealso{
Other results: 
\code{\link{results}()},
\code{\link{results_to_file}()}
}
\concept{results}
