% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.checked_results}
\alias{print.checked_results}
\alias{print.checked_results_check_task_spec}
\alias{print.checked_results_revdep_check_task_spec}
\title{Print checked results}
\usage{
\method{print}{checked_results}(x, ...)

\method{print}{checked_results_check_task_spec}(x, keep = options::opt("results_keep"), ...)

\method{print}{checked_results_revdep_check_task_spec}(x, ...)
}
\arguments{
\item{x}{an object to be printed.}

\item{...}{other parameters.}

\item{keep}{character vector indicating which packages should be included in the results.
"all" means that all packages are kept. If "issues" then only packages
with issues identified, whereas "potential_issues" stands for keeping
packages with both "issues" and "potential_issues". (Defaults to \code{"all"}, overwritable using option 'checked.results_keep' or environment variable 'R_CHECKED_RESULTS_KEEP')}
}
\description{
Print checked results
}
\seealso{
Other results: 
\code{\link{results}()},
\code{\link{results_to_file}()}
}
\concept{results}
