% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{results.check_design}
\title{Check results}
\usage{
results(x, ...)

\method{results}{check_design}(x, error_on = c("never", "issues", "potential_issues"), ...)
}
\arguments{
\item{x}{\code{\link[checked]{check_design}} object.}

\item{...}{other parameters.}

\item{error_on}{character vector indicating whether R error should be thrown
when issues are discovered when generating results. "never" means that no
errors are thrown. If "issues" then errors are emitted only on issues, whereas
"potential issues" stands for error on both issues and potential issues.}
}
\description{
Get R CMD check results
}
