% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{results.check_design}
\title{Check results}
\usage{
results(x, ...)

\method{results}{check_design}(x, error_on = options::opt("error_on"), ...)
}
\arguments{
\item{x}{\code{\link[checked]{check_design}} object.}

\item{...}{other parameters.}

\item{error_on}{character vector indicating whether R error should be thrown when issues
are discovered when generating results. "never" means that no errors
are thrown. If "issues" then errors are emitted only on issues, whereas
"potential issues" stands for error on both issues and potential issues. (Defaults to \code{"never"}, overwritable using option 'checked.error_on' or environment variable 'R_CHECKED_ERROR_ON')}
}
\description{
Get R CMD check results
}
\seealso{
Other results: 
\code{\link{print.checked_results}()},
\code{\link{results_to_file}()}
}
\concept{results}
