% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{results.check_design}
\title{Check results}
\usage{
results(x, ...)

\method{results}{check_design}(x, ...)
}
\arguments{
\item{x}{\code{\link[reversecheck]{check_design}} object.}

\item{...}{other parameters.}
}
\description{
Get R CMD check results
}
