% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{rev_dep_check_tasks_df}
\alias{rev_dep_check_tasks_df}
\title{Build Plan for Checking Reverse Dependencies from Package Source}
\usage{
rev_dep_check_tasks_df(
  path,
  output = tempfile(paste(packageName(), Sys.Date(), sep = "-")),
  ...,
  restore
)
}
\arguments{
\item{path}{A file path to a local package source code directory of the
package whose reverse dependencies are to be checked.}

\item{output}{A file path to a directory in which logs and other check
metadata should be saved. Defaults to a new temporary directory.}

\item{...}{Additional arguments passed to individual
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck}} calls}

\item{restore}{Whether existing logs in \code{output} should be used to try to
restore a previous run. When in an interactive session, \code{NA} will indicate
that you should be prompted to confirm. Otherwise, a logical value
indicates whether to restore from or overwrite the existing logs.}
}
\value{
A check plan
}
\description{
Build Plan for Checking Reverse Dependencies from Package Source
}
\examples{
\dontrun{
pkg_source_dir <- "praise"
rev_dep_check_plan(pkg_source_dir)
}

}
