% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run Reverse-Dependency Checks}
\usage{
run(design, ...)
}
\arguments{
\item{...}{Additional arguments unused}

\item{A}{reverse-dependency plan, or an object coercible into a
plan.}

\item{n}{The number of simultaneous child processes to use.}

\item{reporter}{A reporter to provide progress updates. Will default to the
most expressive command-line reporter given your terminal capabilities.}
}
\description{
Run Reverse-Dependency Checks
}
