% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{Task specification}
\usage{
task(..., .subclass = NULL)
}
\arguments{
\item{...}{parameters passed to downstream constructors.}

\item{.subclass}{Additional subclasses.}
}
\description{
Create task specification list which consists of all the details required
to run specific task.
}
\details{
Tasks can be nested, representing either a singular task, or a set of
related tasks.
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{install_task}()},
\code{\link{meta_task}()}
}
\concept{tasks}
