% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{Task specification}
\usage{
task(alias = NULL, package = NULL, env = NULL)
}
\arguments{
\item{alias}{task alias which also serves as unique identifier of the task.}

\item{package}{\code{\link[checked]{package}} object}

\item{env}{environmental variables to be set in separate process running
specific task.}
}
\description{
Create task specification list which consists of all the details required
to run specific task.
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{custom_install_task}()},
\code{\link{install_task}()},
\code{\link{revdep_check_task}()}
}
\concept{tasks}
