% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{Task specification}
\usage{
task(..., .subclass = NULL)
}
\arguments{
\item{alias}{task alias which also serves as unique identifier of the task.}

\item{package}{\code{\link[checked]{package}} object}

\item{tasks}{Optional additional tasks that are pre-requisites for running
this task.}
}
\description{
Create task specification list which consists of all the details required
to run specific task.
}
\details{
Tasks can be nested, representing either a singular task, or a set of
related tasks.
}
\seealso{
Other tasks: 
\code{\link{check_task}()},
\code{\link{custom_install_task}()},
\code{\link{install_task}()},
\code{\link{revdep_check_task}()}
}
\concept{tasks}
