% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{task_formats}
\alias{task_formats}
\title{task formatter bindings}
\usage{
task_formats(g = NULL, nodes = V(g), task = NULL, tasks = list(task))
}
\description{
This bit of code is intended for use with \code{\link[=cli_task]{cli_task()}}, and allows for us
to layer symbol bindings on top of the environment used for string
interpolation which provide syntactic sugar for common formatting components.
}
