% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{task_graph_create}
\alias{task_graph_create}
\title{Create Task Graph}
\usage{
task_graph_create(df, repos = getOption("repos"))
}
\arguments{
\item{df}{data.frame listing}

\item{repos}{repositories which will be used to identify dependencies chain
to run R CMD checks}
}
\value{
A dependency graph with vertex attributes "root" (a logical value
indicating whether the package as one of the roots used to create the
graph), "status" (installation status) and "order" (installation order).
}
\description{
Create Task Graph
}
