% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{task_graph_neighborhoods}
\alias{task_graph_neighborhoods}
\title{Find Task Neighborhood}
\usage{
task_graph_neighborhoods(g, nodes)
}
\arguments{
\item{g}{A task graph, as produced with \code{\link[=task_graph_create]{task_graph_create()}}}

\item{nodes}{Names or nodes objects of packages whose neighborhoods
should be calculated.}
}
\description{
Find Task Neighborhood
}
