% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{task_graph_neighborhoods}
\alias{task_graph_neighborhoods}
\title{Find Task Neighborhood}
\usage{
task_graph_neighborhoods(g, nodes)
}
\arguments{
\item{g}{A task graph, as produced with \code{\link[=dep_graph_create]{dep_graph_create()}}}

\item{names}{Names of packages whose neighborhoods should be calculcated.}
}
\description{
Find Task Neighborhood
}
